--- a/packages/docs/components/navigation/frontend-components-config-navigation.json
+++ b/packages/docs/components/navigation/frontend-components-config-navigation.json
@@ -3,5 +3,10 @@
 		"title": "Frontend components config",
 		"href": "/frontend-components-config"
 	},
-	"items": []
+	"items": [
+		{
+			"title": "Use cache",
+			"href": "/frontend-components-config/use-cache"
+		}
+	]
 }