--- a/packages/docs/pages/frontend-components-config/index.mdx
+++ b/packages/docs/pages/frontend-components-config/index.mdx
@@ -44,3 +44,7 @@ const { config: webpackConfig, plugins } = config({
 });
 
 ```
+
+## Speed up development builds
+
+You can enable the <a href="/frontend-components-config/use-cache">useCache</a> flag to boost your dev build bootsrap.