### Changes
- offload TS type checking away from the main thread
- enable build cache for builds

### Some testing results

Times are gathered from the RBAC UI

|No changes|Typechecking completely turned off|Typechecking offloaded|First dev build(no cache)|Dev build with cache|
|----------------|-----------------------------------------------|--------------------------------|--------------------------------|---|
|83888 ms|30614 ms|38496 ms|16670 ms + 9287 ms (js + type check)|2678 ms + 5474ms (js + type check)|